+++ packages/@zenithui/day-picker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @zenithui/day-picker
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @zenithui/utils@0.4.0
+
 ## 1.1.0
 
 ### Minor Changes