+++ packages/@zenithui/toast/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @zenithui/toast
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @zenithui/utils@0.4.0
+
 ## 1.0.9
 
 ### Patch Changes